reorganize wifi manager
