+++ include/wifi_manager.h
@@ -1,0 +1,12 @@
+#ifndef WIFI_MANAGER_H
+#define WIFI_MANAGER_H
+
+#include "common.h"
+#include <WiFi.h>
+#include "secrets.h"
+
+
+bool connectToWiFi(unsigned long timeout);
+
+
+#endif // WIFI_MANAGER_H
